Update Products Backend + Test + Frontend

+++ backend/src/main/java/org/example/backend/dto/OrderDTO.java
@@ -4,5 +4,5 @@ import java.util.ArrayList;
 
 public record OrderDTO(
         ArrayList<String> productIds,
-        int price
+        float price
 ) {}
